refactor(db): move repository tests alongside their code
